requirements.txt
asgiref==3.4.1
Django==3.2.9
django-rest-auth==0.9.5
djangorestframework==3.12.4
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
